app/requirements.txt
fastapi
uvicorn
celery
redis
google-generativeai
requests
pytest
python-dotenv
pydantic